# Conflicts:
#	src/main/java/za/ac/cput/domain/ItemType.java
#	src/main/java/za/ac/cput/service/HouseKeepingService.java
#	src/test/java/za/ac/cput/service/HouseKeepingServiceTest.java
#	src/test/java/za/ac/cput/service/SupplierServiceTest.java

+++ src/main/java/za/ac/cput/repository/ManagerRepository.java
@@ -6,7 +6,7 @@ import za.ac.cput.domain.Manager;
 
 
 @Repository
-public interface ManagerRepository extends JpaRepository<Manager, Integer> {
+public interface ManagerRepository extends JpaRepository<Manager, Long> {
 
 }
